Directory restructure after review
