+++ litmus-portal/graphql-server/pkg/graphql/mutations/mutation.go
@@ -1,7 +1,6 @@
 package mutations
 
 import (
-	"fmt"
 	"log"
 	"strconv"
 	"time"
@@ -13,14 +12,15 @@ import (
 	"github.com/jinzhu/copier"
 
 	"github.com/google/uuid"
+	"github.com/pkg/errors"
+	"go.mongodb.org/mongo-driver/bson"
+
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/graph/model"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/cluster"
 	store "github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/data-store"
 	database "github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/database/mongodb"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/pkg/graphql/subscriptions"
 	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/utils"
-	"github.com/pkg/errors"
-	"go.mongodb.org/mongo-driver/bson"
 )
 
 //ClusterRegister creates an entry for a new cluster in DB and generates the url used to apply manifest
@@ -174,304 +174,6 @@ func LogsHandler(podLog model.PodLog, r store.StateData) (string, error) {
 		return "LOGS SENT SUCCESSFULLY", nil
 	}
 	return "LOG REQUEST CANCELLED", nil
-}
-
-func CreateDataSource(datasource *model.DSInput) (*model.DSResponse, error) {
-
-	datasourceStatus := utils.TSDBHealthCheck(datasource.DsURL, datasource.DsType)
-
-	if datasourceStatus == "Active" {
-
-		newDS := database.DataSource{
-			DsID:              uuid.New().String(),
-			DsName:            datasource.DsName,
-			DsType:            datasource.DsType,
-			DsURL:             datasource.DsURL,
-			AccessType:        datasource.AccessType,
-			AuthType:          datasource.AuthType,
-			BasicAuthUsername: datasource.BasicAuthUsername,
-			BasicAuthPassword: datasource.BasicAuthPassword,
-			ScrapeInterval:    datasource.ScrapeInterval,
-			QueryTimeout:      datasource.QueryTimeout,
-			HTTPMethod:        datasource.HTTPMethod,
-			ProjectID:         *datasource.ProjectID,
-			CreatedAt:         strconv.FormatInt(time.Now().Unix(), 10),
-			UpdatedAt:         strconv.FormatInt(time.Now().Unix(), 10),
-		}
-
-		err := database.InsertDataSource(newDS)
-		if err != nil {
-			return nil, err
-		}
-
-		var newDSResponse = model.DSResponse{}
-		_ = copier.Copy(&newDSResponse, &newDS)
-
-		return &newDSResponse, nil
-	} else {
-		return nil, errors.New("Datasource: " + datasourceStatus)
-	}
-}
-
-func CreateDashboard(dashboard *model.CreateDBInput) (string, error) {
-
-	newDashboard := database.DashBoard{
-		DbID:        uuid.New().String(),
-		DbName:      dashboard.DbName,
-		DbType:      dashboard.DbType,
-		DsID:        dashboard.DsID,
-		EndTime:     dashboard.EndTime,
-		StartTime:   dashboard.StartTime,
-		RefreshRate: dashboard.RefreshRate,
-		ClusterID:   dashboard.ClusterID,
-		ProjectID:   dashboard.ProjectID,
-		IsRemoved:   false,
-		CreatedAt:   strconv.FormatInt(time.Now().Unix(), 10),
-		UpdatedAt:   strconv.FormatInt(time.Now().Unix(), 10),
-	}
-
-	var (
-		newPanelGroups = make([]database.PanelGroup, len(dashboard.PanelGroups))
-		newPanels      []*database.Panel
-	)
-
-	for i, panel_group := range dashboard.PanelGroups {
-
-		panelGroupID := uuid.New().String()
-		newPanelGroups[i].PanelGroupID = panelGroupID
-		newPanelGroups[i].PanelGroupName = panel_group.PanelGroupName
-
-		for _, panel := range panel_group.Panels {
-			var newPromQueries []*database.PromQuery
-			copier.Copy(&newPromQueries, &panel.PromQueries)
-
-			var newPanelOptions database.PanelOption
-			copier.Copy(&newPanelOptions, &panel.PanelOptions)
-
-			newPanel := database.Panel{
-				PanelID:      uuid.New().String(),
-				PanelOptions: &newPanelOptions,
-				PanelName:    panel.PanelName,
-				PanelGroupID: panelGroupID,
-				PromQueries:  newPromQueries,
-				IsRemoved:    false,
-				XAxisDown:    panel.XAxisDown,
-				YAxisLeft:    panel.YAxisLeft,
-				YAxisRight:   panel.YAxisRight,
-				Unit:         panel.Unit,
-				CreatedAt:    strconv.FormatInt(time.Now().Unix(), 10),
-				UpdatedAt:    strconv.FormatInt(time.Now().Unix(), 10),
-			}
-
-			newPanels = append(newPanels, &newPanel)
-		}
-	}
-	err := database.InsertPanel(newPanels)
-	if err != nil {
-		return "", errors.Errorf("error on inserting panel data", err)
-	}
-	log.Print("sucessfully inserted prom query into promquery-collection")
-
-	newDashboard.PanelGroups = newPanelGroups
-
-	err = database.InsertDashBoard(newDashboard)
-	if err != nil {
-		return "", errors.Errorf("error on inserting panel data", err)
-	}
-	log.Print("sucessfully inserted dashboard into dashboard-collection")
-
-	return "sucessfully created", nil
-}
-
-func UpdateDataSource(datasource model.DSInput) (*model.DSResponse, error) {
-	timestamp := strconv.FormatInt(time.Now().Unix(), 10)
-
-	if datasource.DsID == nil || *datasource.DsID == "" {
-		return nil, errors.New("Datasource ID is nil or empty")
-	}
-
-	query := bson.D{{"ds_id", datasource.DsID}}
-
-	update := bson.D{{"$set", bson.D{{"ds_name", datasource.DsName},
-		{"ds_url", datasource.DsURL}, {"access_type", datasource.AccessType},
-		{"auth_type", datasource.AuthType}, {"basic_auth_username", datasource.BasicAuthUsername},
-		{"basic_auth_password", datasource.BasicAuthPassword}, {"scrape_interval", datasource.ScrapeInterval},
-		{"query_timeout", datasource.QueryTimeout}, {"http_method", datasource.HTTPMethod},
-		{"updated_at", timestamp}}}}
-
-	err := database.UpdateDataSource(query, update)
-	if err != nil {
-		return nil, err
-	}
-
-	return &model.DSResponse{
-		DsID:              datasource.DsID,
-		DsName:            &datasource.DsName,
-		DsType:            &datasource.DsType,
-		DsURL:             &datasource.DsURL,
-		AccessType:        &datasource.AccessType,
-		AuthType:          &datasource.AuthType,
-		BasicAuthPassword: datasource.BasicAuthPassword,
-		BasicAuthUsername: datasource.BasicAuthUsername,
-		ScrapeInterval:    &datasource.ScrapeInterval,
-		QueryTimeout:      &datasource.QueryTimeout,
-		HTTPMethod:        &datasource.HTTPMethod,
-		UpdatedAt:         &timestamp,
-	}, nil
-}
-
-func UpdateDashBoard(dashboard *model.UpdataDBInput) (string, error) {
-	timestamp := strconv.FormatInt(time.Now().Unix(), 10)
-
-	if dashboard.DbID == "" && dashboard.DsID == "" {
-		return "", errors.New("DashBoard ID or Datasource is nil or empty")
-	}
-
-	query := bson.D{{"db_id", dashboard.DbID}}
-
-	update := bson.D{{"$set", bson.D{{"ds_id", dashboard.DsID}, {"db_name", dashboard.DbName},
-		{"db_type", dashboard.DbType}, {"end_time", dashboard.EndTime},
-		{"start_time", dashboard.StartTime}, {"refresh_rate", dashboard.RefreshRate},
-		{"panel_groups", dashboard.PanelGroups}, {"updated_at", timestamp}}}}
-
-	err := database.UpdateDashboard(query, update)
-	if err != nil {
-		return "", err
-	}
-
-	return "successfully updated", nil
-}
-
-func UpdatePanel(panels []*model.Panel) (string, error) {
-
-	for _, panel := range panels {
-		timestamp := strconv.FormatInt(time.Now().Unix(), 10)
-
-		if *panel.PanelID == "" && *panel.PanelGroupID == "" {
-			return "", errors.New("DashBoard ID or Datasource is nil or empty")
-		}
-
-		var newPanelOption database.PanelOption
-		copier.Copy(&newPanelOption, &panel.PanelOptions)
-
-		var newPromQueries []database.PromQuery
-		copier.Copy(&newPromQueries, panel.PromQueries)
-
-		query := bson.D{{"panel_id", panel.PanelID}}
-
-		update := bson.D{{"$set", bson.D{{"panel_name", panel.PanelName},
-			{"panel_group_id", panel.PanelGroupID}, {"panel_options", newPanelOption},
-			{"prom_queries", newPromQueries}, {"updated_at", timestamp},
-			{"y_axis_left", panel.YAxisLeft}, {"y_axis_right", panel.YAxisRight},
-			{"x_axis_down", panel.XAxisDown}, {"unit", panel.Unit}}}}
-
-		err := database.UpdatePanel(query, update)
-		if err != nil {
-			return "", err
-		}
-	}
-
-	return "successfully updated", nil
-}
-
-func DeleteDashboard(db_id *string) (bool, error) {
-
-	dashboardQuery := bson.M{"db_id": db_id, "is_removed": false}
-	dashboard, err := database.GetDashboard(dashboardQuery)
-	if err != nil {
-		return false, fmt.Errorf("failed to list dashboard, error: %v", err)
-	}
-
-	for _, panelGroup := range dashboard.PanelGroups {
-		listPanelQuery := bson.M{"panel_group_id": panelGroup.PanelGroupID, "is_removed": false}
-		panels, err := database.ListPanel(listPanelQuery)
-		if err != nil {
-			return false, fmt.Errorf("failed to list Panel, error: %v", err)
-		}
-
-		for _, panel := range panels {
-			time := strconv.FormatInt(time.Now().Unix(), 10)
-
-			query := bson.D{{"panel_id", panel.PanelID}, {"is_removed", false}}
-			update := bson.D{{"$set", bson.D{{"is_removed", true}, {"updated_at", time}}}}
-
-			err := database.UpdatePanel(query, update)
-			if err != nil {
-				return false, fmt.Errorf("failed to delete panel, error: %v", err)
-			}
-		}
-	}
-
-	time := strconv.FormatInt(time.Now().Unix(), 10)
-
-	query := bson.D{{"db_id", db_id}}
-	update := bson.D{{"$set", bson.D{{"is_removed", true}, {"updated_at", time}}}}
-
-	err = database.UpdateDashboard(query, update)
-	if err != nil {
-		return false, fmt.Errorf("failed to delete dashboard, error: %v", err)
-	}
-
-	return true, nil
-}
-
-func DeleteDataSource(input model.DeleteDSInput) (bool, error) {
-
-	time := strconv.FormatInt(time.Now().Unix(), 10)
-
-	listDBQuery := bson.M{"ds_id": input.DsID, "is_removed": false}
-	dashboards, err := database.ListDashboard(listDBQuery)
-	if err != nil {
-		return false, fmt.Errorf("failed to list dashboard, error: %v", err)
-	}
-
-	if input.ForceDelete == true {
-		for _, dashboard := range dashboards {
-
-			for _, panelGroup := range dashboard.PanelGroups {
-				listPanelQuery := bson.M{"panel_group_id": panelGroup.PanelGroupID, "is_removed": false}
-				panels, err := database.ListPanel(listPanelQuery)
-				if err != nil {
-					return false, fmt.Errorf("failed to list Panel, error: %v", err)
-				}
-
-				for _, panel := range panels {
-					query := bson.D{{"panel_id", panel.PanelID}, {"is_removed", false}}
-					update := bson.D{{"$set", bson.D{{"is_removed", true}, {"updated_at", time}}}}
-
-					err := database.UpdatePanel(query, update)
-					if err != nil {
-						return false, fmt.Errorf("failed to delete panel, error: %v", err)
-					}
-				}
-			}
-			updateDBQuery := bson.D{{"db_id", dashboard.DbID}}
-			update := bson.D{{"$set", bson.D{{"is_removed", true}, {"updated_at", time}}}}
-
-			err = database.UpdateDashboard(updateDBQuery, update)
-			if err != nil {
-				return false, fmt.Errorf("failed to delete dashboard, error: %v", err)
-			}
-		}
-
-	} else if len(dashboards) > 0 {
-		var db_names []string
-		for _, dashboard := range dashboards {
-			db_names = append(db_names, dashboard.DbName)
-		}
-
-		return false, fmt.Errorf("failed to delete datasource, dashboard(s) are attached to the datasource: %v", db_names)
-	}
-
-	updateDSQuery := bson.D{{"ds_id", input.DsID}}
-	update := bson.D{{"$set", bson.D{{"is_removed", true}, {"updated_at", time}}}}
-
-	err = database.UpdateDataSource(updateDSQuery, update)
-	if err != nil {
-		return false, fmt.Errorf("failed to delete datasource, error: %v", err)
-	}
-
-	return true, nil
 }
 
 func DeleteCluster(cluster_id string, r store.StateData) (string, error) {
